requirements.txt
Django==2.1.2
django-extensions==2.0.7
django-rest-auth==0.9.3
djangorestframework==3.8.2
django-rest-swagger==2.2.0
envparse==0.2.0
psycopg2-binary<2.8
django-db-geventpool
bumpversion<0.6
gunicorn
gevent==1.3.4
requests==2.19.1

# Logging
python-json-logger==0.1.8
structlog==18.1.0

#Database
redis==2.10.6

#Tools
python-masscan==0.1.2
dnspython==1.15.0

# Testing
django-test-plus==1.1.1
factory-boy==2.11.1
pytest==3.6.3
pytest-cov==2.5.1
pytest-django==3.3.2
pytest-sugar==0.9.1
pytest-xdist==1.22.2
whitenoise<3.4.0

mongoengine==0.8.0
